# docs/swagger/admin-api.yaml
openapi: 3.0.3
info:
  title: メダカライブオークション - 管理者API
  description: |
    メダカライブオークションシステムの管理者向けAPI仕様書
    
    ## 認証
    すべてのエンドポイントは`Bearer`トークン認証が必要です。
    また、`admin`ロールを持つユーザーのみアクセス可能です。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: Development server
...
paths:
...
  /admin/auctions/{auctionId}/items:
    get:
      tags:
        - Items
      summary: 出品生体一覧取得
      operationId: getItems
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, approved, rejected, registered, live, sold, unsold, cancelled]
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Item'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    post:
      tags:
        - Items
      summary: 出品生体作成
      operationId: createItem
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: 作成成功
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/auctions/{auctionId}/items/{id}:
    get:
      tags:
        - Items
      summary: 出品生体詳細取得
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      item:
                        $ref: '#/components/schemas/ItemDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Items
      summary: 出品生体更新
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: 更新成功
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags:
        - Items
      summary: 出品生体削除
      operationId: deleteItem
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/auctions/{auctionId}/items/{id}/media:
    post:
      tags:
        - Items
      summary: メディアアップロード
      operationId: uploadMedia
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                media_type:
                  type: string
                  enum: [photo_top, photo_side, photo_other, video_top, video_swim]
      responses:
        '201':
          description: アップロード成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      media:
                        $ref: '#/components/schemas/ItemMedia'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/auctions/{auctionId}/items/{id}/media/{mediaId}:
    delete:
      tags:
        - Items
      summary: メディア削除
      operationId: deleteMedia
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/auctions/{auctionId}/items/{id}/media/{mediaId}/thumbnail:
    patch:
      tags:
        - Items
      summary: サムネイル設定
      operationId: setThumbnail
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 設定成功
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/auctions/{auctionId}/items/bulk-status:
    patch:
      tags:
        - Items
      summary: 一括ステータス更新
      operationId: bulkUpdateStatus
      security:
        - bearerAuth: []
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
                - status
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                status:
                  type: string
                  enum: [approved, rejected, registered]
      responses:
        '200':
          description: 更新成功
        '422':
          $ref: '#/components/responses/ValidationError'

  # ==================== Announcements ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: リソースが見つかりません。
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
  schemas:
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
...
    ItemSummary:
      type: object
      properties:
        id:
          type: integer
        item_number:
          type: integer
        species_name:
          type: string
        current_price:
          type: number
        thumbnail_path:
          type: string
    Item:
      allOf:
        - $ref: '#/components/schemas/ItemSummary'
        - type: object
          properties:
            quantity:
              type: integer
            start_price:
              type: number
            status:
              type: string
            is_premium:
              type: boolean
            seller_name:
              type: string
    ItemDetail:
      allOf:
        - $ref: '#/components/schemas/Item'
        - type: object
          properties:
            estimated_price:
              type: number
            bid_increment:
              type: number
            inspection_info:
              type: string
            individual_info:
              type: string
            notes:
              type: string
            unsold_action:
              type: string
            media:
              type: array
              items:
                $ref: '#/components/schemas/ItemMedia'
    ItemCreate:
      type: object
      required:
        - seller_profile_id
        - species_name
        - quantity
        - start_price
      properties:
        seller_profile_id:
          type: integer
        species_name:
          type: string
        quantity:
          type: integer
        start_price:
          type: number
        estimated_price:
          type: number
        bid_increment:
          type: number
        inspection_info:
          type: string
        individual_info:
          type: string
        notes:
          type: string
        is_premium:
          type: boolean
        unsold_action:
          type: string
          enum: [return, keep, dispose]
    ItemUpdate:
      $ref: '#/components/schemas/ItemCreate'
    ItemMedia:
      type: object
      properties:
        id:
          type: integer
        item_id:
          type: integer
        media_type:
          type: string
        file_path:
          type: string
        thumbnail_path:
          type: string
        is_thumbnail:
          type: boolean
        display_order:
          type: integer
...